working synthetic funcs with gpt-j

+++ mprompt/llm.py
@@ -8,16 +8,19 @@ import pickle as pkl
 import langchain
 from scipy.special import softmax
 from langchain.llms.base import LLM
+import torch
 from langchain.cache import InMemoryCache
 from transformers import AutoConfig, AutoModel, AutoTokenizer, AutoModelForCausalLM
 from transformers import T5Tokenizer, T5ForConditionalGeneration
 langchain.llm_cache = InMemoryCache()
+
 
 def get_llm(checkpoint='openai'):
     if checkpoint.startswith('text-da'):
         return llm_openai()
     else:
         return llm_hf(checkpoint)
+
 
 def llm_openai(checkpoint='text-davinci-003') -> LLM:
     from langchain.llms import OpenAI
@@ -27,12 +30,15 @@ def llm_openai(checkpoint='text-davinci-003') -> LLM:
         # stop='.',
     )
 
+
 def _get_tokenizer(checkpoint):
     if 'facebook/opt' in checkpoint:
-        # opt can't use fast tokenizer...https://huggingface.co/docs/transformers/model_doc/opt
+        # opt can't use fast tokenizer
+        # https://huggingface.co/docs/transformers/model_doc/opt
         return AutoTokenizer.from_pretrained(checkpoint, use_fast=False)
     else:
         return AutoTokenizer.from_pretrained(checkpoint, use_fast=True)
+
 
 def llm_hf(checkpoint='google/flan-t5-xl') -> LLM:
     class LLM_HF(LLM):
@@ -45,7 +51,8 @@ def llm_hf(checkpoint='google/flan-t5-xl') -> LLM:
                 checkpoint, device_map="auto")
         else:
             _model = AutoModelForCausalLM.from_pretrained(
-                checkpoint, device_map="auto")
+                checkpoint, device_map="auto",
+                torch_dtype=torch.float16)
 
         @property
         def _llm_type(self) -> str:
@@ -75,17 +82,21 @@ def llm_hf(checkpoint='google/flan-t5-xl') -> LLM:
             # Get first token id in target_token_str
             target_token_id = self._tokenizer(target_token_str)['input_ids'][0]
 
-            inputs = self._tokenizer(prompt, return_tensors="pt").to(self._model.device)
-            logits = self._model(**inputs)['logits'].detach().cpu()  # shape is (batch_size, seq_len, vocab_size)
-            probs_next_token = softmax(logits[0, -1, :].numpy().flatten())  # shape is (vocab_size,)
-            # token_output_ids = self.tokenizer.convert_tokens_to_ids(target_tokens)
+            # get prob of target token
+            inputs = self._tokenizer(
+                prompt, return_tensors="pt").to(self._model.device)
+            # shape is (batch_size, seq_len, vocab_size)
+            logits = self._model(**inputs)['logits'].detach().cpu()
+            # shape is (vocab_size,)
+            probs_next_token = softmax(logits[0, -1, :].numpy().flatten())
+            # token_output_ids = self._tokenizer.convert_tokens_to_ids(target_tokens)
+            # str(self._tokenizer.convert_ids_to_tokens([np.argmax(probs_next_token)])[0])
             return probs_next_token[target_token_id]
 
         @property
         def _identifying_params(self) -> Mapping[str, Any]:
             """Get the identifying parameters."""
             return vars(self)
-        
 
     return LLM_HF()
 
@@ -102,11 +113,7 @@ if __name__ == '__main__':
     ]:
         target_token = ' Yes.'
         # answer = llm(prompt) # this is weird, adds some kind of asynchrony or something
-        logit_target = llm.get_logit_for_target_token(prompt, target_token=target_token)
+        logit_target = llm.get_logit_for_target_token(
+            prompt, target_token=target_token)
         # print(prompt.strip(), answer, f'logit for logit_target: {logit_target:0.2f}')
         print(repr(prompt), logit_target)
-
-
-
-
-
